Moves Highlighting into Dtu::SolrDocument namespace and tests it
