+++ app/models/dtu/solr_document.rb
@@ -1,0 +1,6 @@
+module Dtu
+  class SolrDocument
+    include Blacklight::Solr::Document
+    include Dtu::SolrDocument::Highlighting
+  end
+end
