+++ dtu_blacklight_common.gemspec
@@ -22,7 +22,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'execjs'
   spec.add_development_dependency "sqlite3"
   spec.add_development_dependency "rspec-rails"
-  spec.add_development_dependency "rspec-core"
+  spec.add_development_dependency "capybara"
   spec.add_development_dependency "engine_cart"
   spec.add_dependency "blacklight", '5.7'
 end
