+++ spec/models/concerns/solr_document/highlighting_spec.rb
@@ -1,14 +1,13 @@
-describe Highlighting do
-
-  before do
-    class HighlightingTestClass
-      include Highlighting
-
+describe Dtu::SolrDocument::Highlighting do
+  let(:described_class) do
+    Class.new do
+        include Blacklight::Solr::Document
+        include Dtu::SolrDocument::Highlighting
     end
   end
 
   let(:solr_response) { { 'highlighting' => highlighting_values } }
-  let(:document) { HighlightingTestClass.new({ id: '2059', cluster_id_ss: '598732111' }, solr_response ) }
+  let(:document) { described_class.new({ id: '2059', cluster_id_ss: '598732111' }, solr_response ) }
   let(:highlighting_values) { {"2059"=>{"abstract_ts"=>["highlight value 1"]}} }
 
   describe 'has_highlight_field?' do
